Add /gameover notification API
The stastitics service need to get notified when a game ends, so it can
update the players statistics.

The /gameover API will perform a longpoll that returns the list of
new ended games since last call (using the `since` parameter). Initial
call may return the whole list of games stored on the server... Which in
turn is gonna trigger a major re-computation of all statistics of all
players. How long will that take? We will see in production, however the
upside of that is that it's gonna be easy to change the ranking
algorithm (for example), restart the statistics module, it will do a
full update. If we make sure to restart those modules during the night,
this shouldn't be any problem.

This API is protected using API_SECRET (needs to be specified when
deploying the coordinator service).