Make each test self sufficient
I wanted to add new notifications-related tests, but got stuck by the
fact that existing tests where dependant on what was run by the previous
tests.

So I made all tests independant from the others. To do so, a new server
and a new DB is created between each test.  `before` and `after` are
turned into `beforeEach` and `afterEach`.